End of Permission

+++ client/src/app/Master/size/size-alloc-card/size-alloc-card.component.ts
@@ -12,7 +12,7 @@ import { MasterService } from '_services/master.service';
 @Component({
   selector: 'app-size-alloc-card',
   templateUrl: './size-alloc-card.component.html',
-  styleUrls: ['./size-alloc-card.component.css']
+  styleUrls: ['./size-alloc-card.component.css'],
 })
 export class SizeAllocCardComponent implements OnInit {
   sizeAllocForm: FormGroup;
@@ -25,18 +25,22 @@ export class SizeAllocCardComponent implements OnInit {
   removeButton: boolean = false;
   saveButton: boolean = false;
   validationErrors: string[] = [];
-  
+
   public col: IgxColumnComponent;
   public pWidth: string;
   public nWidth: string;
-  
-  @ViewChild("aSizeGrid", { static: true }) 
+
+  @ViewChild('aSizeGrid', { static: true })
   public aSizeGrid: IgxGridComponent;
-  @ViewChild("naSizeGrid", { static: true }) 
+  @ViewChild('naSizeGrid', { static: true })
   public naSizeGrid: IgxGridComponent;
 
-  constructor(private accountService: AccountService, private fb: FormBuilder
-    ,private masterService: MasterService ,private toastr: ToastrService) { }
+  constructor(
+    private accountService: AccountService,
+    private fb: FormBuilder,
+    private masterService: MasterService,
+    private toastr: ToastrService
+  ) {}
 
   ngOnInit(): void {
     this.initilizeForm();
@@ -45,17 +49,17 @@ export class SizeAllocCardComponent implements OnInit {
   }
 
   initilizeForm() {
-    this.accountService.currentUser$.forEach(element => {
+    this.accountService.currentUser$.forEach((element) => {
       this.user = element;
-      });
+    });
 
-    this.sizeAllocForm = this.fb.group ({
-      autoId : [0],
-      createUserId : this.user.userId,
-      sizeCard: ['',Validators.required],
+    this.sizeAllocForm = this.fb.group({
+      autoId: [0],
+      createUserId: this.user.userId,
+      sizeCard: ['', Validators.required],
       // color:['', Validators.required]
-    })
-  } 
+    });
+  }
 
   public singleSelection(event: IComboSelectionChangeEventArgs) {
     if (event.added.length) {
@@ -70,75 +74,76 @@ export class SizeAllocCardComponent implements OnInit {
   }
 
   //// loads Size cards
-  loadSizeCard(){
-    this.masterService.getSizeCard().subscribe(result => {
+  loadSizeCard() {
+    this.masterService.getSizeCard().subscribe((result) => {
       this.sizeCardList = result;
-    })
+    });
   }
 
   onSizeCardSelect(event) {
     this.clearGridRows();
-    for(const item of event.added) {
+    for (const item of event.added) {
       this.loadSizeList(item);
     }
   }
 
-  getButtonPermission() {    
-    // this.userObj = JSON.parse(localStorage.getItem('user'));  
-    this.authMenus = this.user.permitMenus;  
-    
-    if(this.authMenus != null) {
-    if(this.authMenus.filter(x => x.menuName == 'SaveSizeAllocation' && x.mType == 'B').length > 0) {
-      this.saveButton = true;
+  getButtonPermission() {
+    // this.userObj = JSON.parse(localStorage.getItem('user'));
+    this.authMenus = this.user.permitMenus;
+
+    if (this.authMenus != null) {
+      if (this.authMenus.filter((x) => x.autoIdx == 39).length > 0) {
+        this.saveButton = true;
+      }
+      if (this.authMenus.filter((x) => x.autoIdx == 41).length > 0) {
+        this.removeButton = true;
+      }
     }
-    if(this.authMenus.filter(x => x.menuName == 'DeleteSizeAllocation' && x.mType == 'B').length > 0) {
-      this.removeButton = true;
-    }
-  }
-    // console.log(this.authMenus);
-    // console.log(this.saveButton);
-    // console.log(this.removeButton);
   }
 
-  //// LOADS ASSIGN AND UNASIGNED SIZE LIST 
-  loadSizeList (sizeCardId) {
-    this.masterService.getSizeAllocDetails(sizeCardId).subscribe(result => {
-      if(result.length > 0) {
-        this.aSizeList = result.filter(x => x.isAsign == true);
-        this.naSizeList = result.filter(x => x.isAsign == false);
-      }      
-    })
+  //// LOADS ASSIGN AND UNASIGNED SIZE LIST
+  loadSizeList(sizeCardId) {
+    this.masterService.getSizeAllocDetails(sizeCardId).subscribe((result) => {
+      if (result.length > 0) {
+        this.aSizeList = result.filter((x) => x.isAsign == true);
+        this.naSizeList = result.filter((x) => x.isAsign == false);
+      }
+    });
   }
 
   saveSizeAllocation() {
-    var selectedRows = this.naSizeGrid.selectedRows;
-    var sizeCardId = this.sizeAllocForm.get("sizeCard").value[0];
-    var sizeList =[];
+    if (this.saveButton == true) {
+      var selectedRows = this.naSizeGrid.selectedRows;
+      var sizeCardId = this.sizeAllocForm.get('sizeCard').value[0];
+      var sizeList = [];
 
-    selectedRows.forEach(menuId => {      
-      var data = {
-        sizeCardId: sizeCardId,
-        sizeId: menuId,
-        createUserId: this.user.userId,
-      };
+      selectedRows.forEach((menuId) => {
+        var data = {
+          sizeCardId: sizeCardId,
+          sizeId: menuId,
+          createUserId: this.user.userId,
+        };
 
-      sizeList.push(data);
-    });
+        sizeList.push(data);
+      });
 
-    console.log(sizeList);
-    this.masterService.saveSizeAllocation(sizeList).subscribe((result) =>{
-      if(result == 1) {
-        this.toastr.success("Size Assign Successfully !!!");
-        this.clearGridRows();
-        this.loadSizeList(sizeCardId);
-      } else if (result == -1) {
-        this.toastr.warning("Size Assign failed !!!");
-        this.clearGridRows();
-        this.loadSizeList(sizeCardId);
-      } else {
-        this.toastr.warning("Contact Admin. Error No:- " + result.toString());
-      }
-    })
+      // console.log(sizeList);
+      this.masterService.saveSizeAllocation(sizeList).subscribe((result) => {
+        if (result == 1) {
+          this.toastr.success('Size Assign Successfully !!!');
+          this.clearGridRows();
+          this.loadSizeList(sizeCardId);
+        } else if (result == -1) {
+          this.toastr.warning('Size Assign failed !!!');
+          this.clearGridRows();
+          this.loadSizeList(sizeCardId);
+        } else {
+          this.toastr.warning('Contact Admin. Error No:- ' + result.toString());
+        }
+      });
+    } else {
+      this.toastr.error('Save Permission denied !!!');
+    }
   }
 
   clearGridRows() {
@@ -149,36 +154,37 @@ export class SizeAllocCardComponent implements OnInit {
   }
 
   deleteSizeAllocation() {
-    var selectedRows = this.naSizeGrid.selectedRows;
-    var sizeCardId = this.sizeAllocForm.get("sizeCard").value[0];
-    var sizeList =[];
+    if (this.removeButton == true) {
+      var selectedRows = this.naSizeGrid.selectedRows;
+      var sizeCardId = this.sizeAllocForm.get('sizeCard').value[0];
+      var sizeList = [];
 
-    selectedRows.forEach(menuId => {      
-      var data = {
-        sizeCardId: sizeCardId,
-        sizeId: menuId,
-        createUserId: this.user.userId,
-      };
+      selectedRows.forEach((menuId) => {
+        var data = {
+          sizeCardId: sizeCardId,
+          sizeId: menuId,
+          createUserId: this.user.userId,
+        };
 
-      sizeList.push(data);
-    });
+        sizeList.push(data);
+      });
 
-    // console.log(colorList);
-    this.masterService.deleteSizeAllocation(sizeList).subscribe((result) =>{
-      if(result == 1) {
-        this.toastr.success("Size remove Successfully !!!");
-        this.clearGridRows();
-        this.loadSizeList(sizeCardId);
-      } else if (result == -1) {
-        this.toastr.warning("Size remove failed !!!");
-        this.clearGridRows();
-        this.loadSizeList(sizeCardId);
-      } else {
-        this.toastr.warning("Contact Admin. Error No:- " + result.toString());
-      }
-    })
+      // console.log(colorList);
+      this.masterService.deleteSizeAllocation(sizeList).subscribe((result) => {
+        if (result == 1) {
+          this.toastr.success('Size remove Successfully !!!');
+          this.clearGridRows();
+          this.loadSizeList(sizeCardId);
+        } else if (result == -1) {
+          this.toastr.warning('Size remove failed !!!');
+          this.clearGridRows();
+          this.loadSizeList(sizeCardId);
+        } else {
+          this.toastr.warning('Contact Admin. Error No:- ' + result.toString());
+        }
+      });
+    } else {
+      this.toastr.error('Delete permission denied !!!');
+    }
   }
-
-  
-
 }
